test: move `deployProxy(...)` function inside `fixture.ts`
